requirements.txt
# PyTorch CUDA 12.6
--extra-index-url https://download.pytorch.org/whl/cu126
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0

# spconv под CUDA 12.6
spconv-cu126

# Общие утилиты
tensorboard
dropblock
opencv-python
tqdm
pyaml
open3d
pypcd4
imageio
matplotlib
Pillow
rerun-sdk==0.27.2
numpy
scipy==1.15.3
GitPython
docker

# PyG (PyTorch Geometric)
-f https://data.pyg.org/whl/torch-2.8.0+cu126.html
pyg_lib==0.5.0+pt28cu126
torch_scatter==2.1.2+pt28cu126
torch_sparse==0.6.18+pt28cu126
torch_cluster==1.6.3+pt28cu126
torch_spline_conv==1.2.2+pt28cu126
